refactor: Result enum 패키지 변경
