+++ src/main/java/blackjack/domain/card/Bust.java
@@ -1,7 +1,5 @@
 package blackjack.domain.card;
 
-import blackjack.domain.player.Result;
-
 public class Bust extends Finished {
     Bust(final Cards cards) {
         super(cards);
