+++ src/main/java/blackjack/domain/card/Hand.java
@@ -1,7 +1,5 @@
 package blackjack.domain.card;
 
-import blackjack.domain.player.Result;
-
 public interface Hand {
 
     Hand draw(final Card card);
