+++ src/main/java/blackjack/domain/card/Running.java
@@ -1,7 +1,5 @@
 package blackjack.domain.card;
 
-import blackjack.domain.player.Result;
-
 public abstract class Running extends AbstractHand {
     Running(final Cards cards) {
         super(cards);
