+++ src/test/java/blackjack/domain/game/BetsTest.java
@@ -3,8 +3,8 @@ package blackjack.domain.game;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.entry;
 
+import blackjack.domain.card.Result;
 import blackjack.domain.player.Name;
-import blackjack.domain.player.Result;
 import java.util.Map;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
