+++ src/test/java/blackjack/domain/game/MoneyTest.java
@@ -4,7 +4,7 @@ package blackjack.domain.game;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import blackjack.domain.player.Result;
+import blackjack.domain.card.Result;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
